refactor(server-mongo-testing): extract interface from MongoDbRule
